fix(runtime): align timestamp_ms with detection-v1 contract (int)
- Frame.timestamp_ms: float -> int
- CameraState.last_timestamp_ms: float -> int
- WebcamVideoSource: int(time.monotonic() * 1000)
- Test: assert isinstance(frame.timestamp_ms, int)

+++ integration/video/sources.py
@@ -133,7 +133,7 @@ class WebcamVideoSource(VideoSource):
         if not ret:
             return None
         h, w = bgr.shape[:2]
-        timestamp_ms = time.monotonic() * 1000
+        timestamp_ms = int(time.monotonic() * 1000)
         self.state.record_frame(timestamp_ms)
         return Frame(
             camera_id=self.camera_id,
